(split)bug: 75153 jira: HS-5126
Integrating from OCTOPUS to main

Widen ParsedDocument.size field from int to long.  Integer type was causing document size >= 2GB to get truncated and become negative.  This would eventually cause SQLException during insertion to mail_item table.

Redolog version is bumped up to 1.42 to change CreateMessage.size from int to long.

http://bugzilla.zimbra.com/show_bug.cgi?id=75153

Copied from Perforce
 Change: 376225

+++ src/java/com/zimbra/cs/mailbox/Mailbox.java
@@ -5080,7 +5080,7 @@ public class Mailbox {
         }
     }
 
-    public Message addMessage(OperationContext octxt, InputStream in, int sizeHint, Long receivedDate, DeliveryOptions dopt, DeliveryContext dctxt)
+    public Message addMessage(OperationContext octxt, InputStream in, long sizeHint, Long receivedDate, DeliveryOptions dopt, DeliveryContext dctxt)
     throws IOException, ServiceException {
         int bufLen = Provisioning.getInstance().getLocalServer().getMailDiskStreamingThreshold();
         CopyInputStream cs = new CopyInputStream(in, sizeHint, bufLen, bufLen);
