+++ src/java/com/zimbra/cs/mime/ParsedMessageOptions.java
@@ -2,12 +2,12 @@
  * ***** BEGIN LICENSE BLOCK *****
  * Zimbra Collaboration Suite Server
  * Copyright (C) 2009, 2010 Zimbra, Inc.
- * 
+ *
  * The contents of this file are subject to the Zimbra Public License
  * Version 1.3 ("License"); you may not use this file except in
  * compliance with the License.  You may obtain a copy of the License at
  * http://www.zimbra.com/license.
- * 
+ *
  * Software distributed under the License is distributed on an "AS IS"
  * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied.
  * ***** END LICENSE BLOCK *****
@@ -28,13 +28,13 @@ public class ParsedMessageOptions {
     private byte[] mRawData;
     private File mFile;
     private String mRawDigest;
-    private Integer mRawSize;
+    private Long mRawSize;
     private Long mReceivedDate;
     private Boolean mIndexAttachments;
-    
+
     public ParsedMessageOptions() {
     }
-    
+
     public ParsedMessageOptions(Blob blob, byte buffer[]) throws IOException {
         this(blob, buffer, null, null);
     }
@@ -51,7 +51,7 @@ public class ParsedMessageOptions {
         if (indexAttachments != null)
             setAttachmentIndexing(indexAttachments);
     }
-    
+
     public ParsedMessageOptions setContent(MimeMessage mimeMessage) {
         if (mRawData != null || mFile != null) {
             throw new IllegalArgumentException("Content can only come from one source.");
@@ -59,7 +59,7 @@ public class ParsedMessageOptions {
         mMimeMessage = mimeMessage;
         return this;
     }
-    
+
     public ParsedMessageOptions setContent(byte[] rawData) {
         if (mMimeMessage != null || mFile != null) {
             throw new IllegalArgumentException("Content can only come from one source.");
@@ -67,7 +67,7 @@ public class ParsedMessageOptions {
         mRawData = rawData;
         return this;
     }
-    
+
     public ParsedMessageOptions setContent(File file) {
         if (mRawData != null || mMimeMessage != null) {
             throw new IllegalArgumentException("Content can only come from one source.");
@@ -75,32 +75,32 @@ public class ParsedMessageOptions {
         mFile = file;
         return this;
     }
-    
+
     public ParsedMessageOptions setDigest(String digest) {
         mRawDigest = digest;
         return this;
     }
-    
+
     public ParsedMessageOptions setSize(long size) {
-        mRawSize = (int) size;
+        mRawSize = size;
         return this;
     }
-    
+
     public ParsedMessageOptions setReceivedDate(long receivedDate) {
         mReceivedDate = receivedDate;
         return this;
     }
-    
+
     public ParsedMessageOptions setAttachmentIndexing(boolean enabled) {
         mIndexAttachments = enabled;
         return this;
     }
-    
+
     public MimeMessage getMimeMessage() { return mMimeMessage; }
     public byte[] getRawData() { return mRawData; }
     public File getFile() { return mFile; }
     public String getDigest() { return mRawDigest; }
-    public Integer getSize() { return mRawSize; }
+    public Long getSize() { return mRawSize; }
     public Long getReceivedDate() { return mReceivedDate; }
     public Boolean getAttachmentIndexing() { return mIndexAttachments; }
 }
